improve testing and consider the best way to address send races
There's a race between sender drop and subscriber receive. If we have a
cache, then we can probably recover the value from the cache, but this
is also racy.

The current approach is a bit tighter, but could be improved.

Testing is improved with the new mock and I tightened up the API
requirements again after reflecting that modifying your deduplicate
under a shared reference was probably a bad idea.