fix the race condition caused by sending values without locking
This fixes the race that could occur between sending values after work
completes and subscribing to the result. Because we didn't send under
the lock a race could occur with subscribing, which causes issues.

I'm fairly sure this is now resolved and so the complicated recovery
logic from cache snooping can be removed.