Make some small changes based on using the crate in a real setting
 - Clean up the generic retrieval function definition (remove Clone)
 - Replace std Mutex with tokio Mutex
 - Add some sample code to the docs about storing a Deduplicate in
   a non-generic struct.

The main fix here is replacing std Mutex with tokio Mutex so that using
deduplicate in an async context will work.

I also removed the requirement to make the retrieval function be Clone.

It took me a little while to figure out how to hold a Deduplicate in a
non-generic struct, so I figured it might be useful to document that.