update lru and acknowledge that interior mutability is good
I've wobbled back and forth on whether mutating operations should
require &mut self. In many use-cases, specifically sharing a Deduplicate
betweeen tasks/threads under Arc, this is an onerous requirement since
it requires a Mutex (or such like) above the Deduplicate to provide
this.

So, let's remove that requirement to make life simpler.

Also: a new version of lru has been released, so let's bump to that.